refactor: delete useless props and move context on root
